Move `dialoguer` and `on_call` modules to `admin` module
